a few tweaks

+++ off-chain/src/main/java/org/cardanofoundation/hydrapoc/generator/RandomUtil.java
@@ -2,7 +2,7 @@ package org.cardanofoundation.hydrapoc.generator;
 
 public class RandomUtil {
 
-    public static long getRandomNumber(long min, long max) {
-        return (long) ((Math.random() * (max - min)) + min);
+    public static int getRandomNumber(int min, int max) {
+        return (int) ((Math.random() * (max - min)) + min);
     }
 }
